klee-verifier: fix memory safety, UB, and design issues from code review
- Replace per-call-site abort rewriting with declaration-body approach
  (fixes malformed IR from double terminators, handles indirect calls)
- Match abort-like functions by explicit name list + noreturn attribute
  instead of overly broad substring match on "abort"
- Guard inf.pointee for null and unsized types before getTypeAllocSize
- Re-fetch target_fn after linkModules to avoid stale pointer
- Fix dangling Twine references by materializing to std::string
- Add assert guards on GetInsertBlock() before ->getParent() chains
- Convert typeContainsPointer from recursive to iterative (worklist)
- Fix findMatchingBracket underflow with unsigned depth counter
- Add empty-string guards before stoul/stoll/stoull calls
- Add range check on stoul result to prevent unsigned truncation
- Remove const from cl::opt (conflicts with LLVM mutable internals)
- Replace heap-address type mangling with deterministic counters
- Warn on postconditions targeting by-value arguments (no-op)
- Scan for narrow inttoptr/ptrtoint after architecture retargeting
- Document pipeline ordering dependency (abort redirect before stub)
- Reorder includes: self → std → llvm → patchestry → local
